Fixed comments, + separate folder servicetests
