+++ Streetcode/Streetcode.XIntegrationTest/ServiceTests/BlobServiceTests/SaveFileTests/SaveFileInStorageBase64Tests.cs
@@ -1,9 +1,9 @@
 ﻿using Newtonsoft.Json;
 using Streetcode.DAL.Entities.Media.Images;
-using Streetcode.XIntegrationTest.BlobServiceTests.Utils;
+using Streetcode.XIntegrationTest.ServiceTests.BlobServiceTests.Utils;
 using Xunit;
 
-namespace Streetcode.XIntegrationTest.BlobServiceTests
+namespace Streetcode.XIntegrationTest.ServiceTests.BlobServiceTests.SaveFileTests
 {
     public class SaveFileInStorageBase64Tests : BlobServiceTestBase
     {
@@ -12,7 +12,7 @@ namespace Streetcode.XIntegrationTest.BlobServiceTests
         }
 
         [Theory]
-        [InlineData("../../../BlobServiceTests/Utils/testData.json")]
+        [InlineData("../../../ServiceTests/BlobServiceTests/Utils/testData.json")]
         public void ShouldSaveFileToBlobStorageBase64_AllValid(string testDataFilePath)
         {
             // Arrange
